指摘対応 - bundle update を実行し、Gemfile.lock を最新化 - horse_scraper.rb を services/scraper 配下に移動 - database.yml を修正し、test環境のDBをdevelop環境と同様のmysqlに変更
